@@ -5,11 +5,13 @@
 
 from pathlib import Path
 
-import reversion_compare
-from poetry_publish.tests.test_project_setup import test_assert_rst_readme as assert_rst_readme
+from creole.setup_utils import update_rst_readme
 from poetry_publish.tests.test_project_setup import test_poetry_check as assert_poetry_check
 from poetry_publish.tests.test_project_setup import test_version as assert_version
+
+import reversion_compare
 from reversion_compare import __version__
+
 
 PACKAGE_ROOT = Path(reversion_compare.__file__).parent.parent
 
@@ -22,13 +24,15 @@ def test_version():
     assert_version(package_root=PACKAGE_ROOT, version=__version__)
 
 
-def test_assert_rst_readme():
-    """
-    Check if own README.rst is up-to-date with README.creole
-    """
-    assert_rst_readme(
-        package_root=PACKAGE_ROOT, version=__version__, filename='README.creole'
+def test_update_rst_readme(capsys):
+    rest_readme_path = update_rst_readme(
+        package_root=PACKAGE_ROOT, filename='README.creole'
     )
+    captured = capsys.readouterr()
+    assert captured.out == 'Generate README.rst from README.creole...nothing changed, ok.\n'
+    assert captured.err == ''
+    assert isinstance(rest_readme_path, Path)
+    assert str(rest_readme_path).endswith('/README.rst')
 
 
 def test_poetry_check():
