@@ -18,15 +18,12 @@
 from django.apps import apps
 from django.contrib.auth.models import User
 from django.test import TestCase
-
+from django_tools.unittest_utils.BrowserDebug import debug_response
 from reversion import get_registered_models
 from reversion.models import Revision, Version
 
-# https://github.com/jedie/django-tools
-from django_tools.unittest_utils.BrowserDebug import debug_response
-
 # Needs to import admin module to register all models via CompareVersionAdmin/VersionAdmin
-from .fixtures import Fixtures
+from reversion_compare_tests.utils.fixtures import Fixtures
 
 
 class BaseTestCase(TestCase):
