@@ -1,6 +1,5 @@
 from reversion_compare.views import HistoryCompareDetailView
-
-from .models import SimpleModel
+from reversion_compare_tests.models import SimpleModel
 
 
 class SimpleModelHistoryCompareView(HistoryCompareDetailView):
